refactor(model): refactor cache
