+++ model/src/cache/discord/process/mod.rs
@@ -1,0 +1,4 @@
+//! Incoming events processing.
+
+pub mod event;
+pub mod resource;
