+++ model/src/cache/model/mod.rs
@@ -1,19 +1,7 @@
 //! Cache models.
 //!
-//! This module contains models used by the cache. These models are based on
-//! [`twilight_model`] models but without unnecessary fields to decrease memory
-//! usage.
-//!
-//! Every model implement the [`Serialize`] and [`Deserialize`] traits.
-//!
-//! [`Serialize`]: serde::Serialize
-//! [`Deserialize`]: serde::Deserialize
+//! This module contains models of cached values. It does not contains cached
+//! Discord objects, which are in the [`discord`] module.
 
 pub mod interaction;
 pub mod message;
-
-mod channel;
-mod guild;
-
-pub use channel::CachedChannel;
-pub use guild::{CachedGuild, CachedRole, CurrentMember};
