--- model/src/cache/process/mod.rs
@@ -1,6 +1,0 @@
-//! Incoming events processing.
-
-mod event;
-mod resource;
-
-pub use event::UpdateCache;
