+++ raidprotect/src/cluster.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 use anyhow::Context;
 use futures::StreamExt;
 use raidprotect_model::{
-    cache::{http::CacheHttp, RedisClient},
+    cache::{discord::http::CacheHttp, CacheClient},
     config::BotConfig,
     database::DbClient,
 };
@@ -59,7 +59,7 @@ impl ShardCluster {
 
         info!("logged as {} with ID {}", application.name, current_user);
 
-        let redis = RedisClient::new(&config.database.redis_uri).await?;
+        let redis = CacheClient::connect(&config.database.redis_uri).await?;
         redis.ping().await.context("failed to connect to redis")?;
 
         let mongodb = DbClient::connect(
@@ -154,7 +154,7 @@ fn presence() -> UpdatePresencePayload {
 #[derive(Debug)]
 pub struct ClusterState {
     /// In-memory cache
-    redis: RedisClient,
+    redis: CacheClient,
     /// MongoDB client
     mongodb: DbClient,
     /// Http client
@@ -166,7 +166,7 @@ pub struct ClusterState {
 impl ClusterState {
     /// Initialize a new [`ClusterState`].
     pub fn new(
-        redis: RedisClient,
+        redis: CacheClient,
         mongodb: DbClient,
         http: Arc<HttpClient>,
         current_user: Id<ApplicationMarker>,
@@ -180,7 +180,7 @@ impl ClusterState {
     }
 
     /// Get the cluster [`RedisClient`].
-    pub fn redis(&self) -> &RedisClient {
+    pub fn redis(&self) -> &CacheClient {
         &self.redis
     }
